refactor: Separate each classes to new files
