+++ VendingMachineApp/VendingMachineApp/Base.lproj/Main.storyboard
@@ -66,7 +66,7 @@
                                                                         </subviews>
                                                                     </stackView>
                                                                     <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="coke" translatesAutoresizingMaskIntoConstraints="NO" id="mjV-Sy-WRw">
-                                                                        <rect key="frame" x="0.0" y="33.5" width="336.5" height="185.50000000000011"/>
+                                                                        <rect key="frame" x="0.0" y="33.5" width="336.5" height="185.5"/>
                                                                     </imageView>
                                                                 </subviews>
                                                             </stackView>
