+++ VendingMachineApp/VendingMachineApp/Views/BeverageButtons/CokeButton.swift
@@ -1,0 +1,15 @@
+//
+//  CokeButton.swift
+//  VendingMachineApp
+//
+//  Created by Chaewan Park on 2020/03/16.
+//  Copyright © 2020 Chaewan Park. All rights reserved.
+//
+
+import UIKit
+
+final class CokeButton: BeverageButton {
+    override func invokeAction(sender: BeverageButton) {
+        action?(Coke())
+    }
+}
