+++ VendingMachineApp/VendingMachineApp/Views/BeverageButtons/GeorgiaButton.swift
@@ -1,0 +1,15 @@
+//
+//  GeorgiaButton.swift
+//  VendingMachineApp
+//
+//  Created by Chaewan Park on 2020/03/16.
+//  Copyright © 2020 Chaewan Park. All rights reserved.
+//
+
+import UIKit
+
+final class GeorgiaButton: BeverageButton {
+    override func invokeAction(sender: BeverageButton) {
+        action?(Georgia())
+    }
+}
